+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to
 ### Added
 
 - Add `appendix` field on `ContractDefinition` model
-- Bind `appendix` field to `AdminContractDefinitionSerializer`
+- Allow to edit `appendix` `ContractDefinition` field through the back office
 
 ## [2.8.0] - 2024-10-16